+++ NEWS.md
@@ -1,7 +1,8 @@
-# TreeSearch 0.3.2.9004 (develoment)
+# TreeSearch 0.3.2.9005 (develoment)
 
 ## New features
  - `ClusteringInfo` Adopts Meila's (2007) clustering information as a tree distance measure (but is not yet production-ready)
+ - Normalization of certain tree distance measures
  - Implement another information theoretic tree distance measure (Smith, in prep)
  
 ## Bug fixes